Add initial properties view
The properties view will replace the old ship and squad
editor for a more generic method that can be used between
the X-Wing ship editor and the Imperial Assault Deployments.
By selecting an item in the TreeView, the property view for
that node will be updated and the editable fields will be
displayed.  When a property is updated it can fire off a
job to generate the new squad data.