Issue #54 Make sure Numeric Fields are Editable
To address the issue of Numeric fields not being editable
create a NumericCellEditor that can be used and a
NumericPropertyDescriptor.   This will correctly allow
all numeric fields to be editable and will return Integer
values for the inputs.  For this application that is all that
is necessary, but could easily be expanded to Floats, Doubles,
and other integer formats.