Refactor for more General Use
Refactored the General items like Players into their
own general use section.  Removed the Dice section from
the general page as it was X-Wing Specific and a pain
to work with from a data entry stand point.

Also separated output directory for player files in the
general section.  Moved Timer as a sub page under the general
section.

Added a field for tracking whether the refresh javascript needs
to be included for Imperial Assault templates.  Should add
same thing to X-Wing to simplify the maintenance.

Updated various jobs to use the new fields and preferences.